+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 ### Changed
 - chk_rep_cfg:  Remove innodb_support_xa from replication config for MySQL 8.0 and above.
 - config/mysql_cfg.py.TEMPLATE:  Add SSL configuration entries.
+- Removed unnecessary \*\*kwargs in function argument list.
 - chk_rep:  Replaced manual creation of MasterRep class instance with mysql_libs.create_instance.
 - run_program, chk_rep, chk_rep_cfg:  Replaced cmds_gen.disconnect with mysql_libs.disconnect.
 - Documentation updates.